requirements.txt
brightdata-sdk
python-dotenv
